挙手判定のプログラムで画像を表示することができた
