--- recognition_startup/readme.md
@@ -1,9 +1,0 @@
-# Overview
-このパッケージはrecognitionパッケージを使うにあたって必要なライブラリなどを一式ダウンロードするものです。
-
-# setup.py
-## 概要
-このプログラムはPC本体にRecognitionで必要になる外部パッケージを一式ダウンロードするシステムです。
-`config`フォルダーに入っている二行目以降のパッケージを探し出してインストールします。
-
-## 関数の紹介
